% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{balance}
\alias{balance}
\title{Compute a simple balance}
\usage{
balance(exports, imports)
}
\arguments{
\item{exports}{Numeric vector of export values.}

\item{imports}{Numeric vector of import values.}
}
\value{
Numeric vector of exports minus imports.
}
\description{
Compute a simple balance
}
