% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-balance_of_trade.R
\name{geom_balance_of_trade}
\alias{geom_balance_of_trade}
\title{geom_balance_of_trade}
\usage{
geom_balance_of_trade(
  data = NULL,
  mapping = NULL,
  stat = StatBalanceOfTrade,
  geom = GeomBalanceOfTrade,
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA
)
}
\arguments{
\item{data}{A data frame (defaults to the plot-level data)}

\item{mapping}{Set of aesthetic mappings, see \code{\link[ggplot2]{geom_ribbon}}}

\item{stat}{Stat to use (defaults to \code{StatBalanceOfTrade})}

\item{geom}{Geom to use (defaults to \code{GeomBalanceOfTrade})}

\item{position}{Position adjustment}

\item{...}{Other arguments passed on to the layer (e.g. \code{na.rm})}

\item{na.rm}{If \code{TRUE}, remove \code{NA}s}

\item{show.legend}{Should this layer be included in the legend?}
}
\value{
A ggplot2 layer representing the balance of trade. Within the layer with be a colored ribbon and two lines and point sets representing your imports and exports
}
\description{
\code{geom_balance_of_trade()} creates a custom ggplot2 ribbon that visualizes the balance of trade by shading the area between exports and imports.
}
\note{
If you map \code{exports}/\code{imports} in the global \code{aes()}, you'll see:
\"The following aesthetics were dropped…exports and imports.\"
This simply means the stat used them to compute the ribbon and then removed
them—your plot still renders correctly.
}
