% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-balance_of_trade.R
\docType{data}
\name{GeomBalanceOfTrade}
\alias{GeomBalanceOfTrade}
\title{GeomBalanceOfTrade}
\format{
An object of class \code{GeomBalanceOfTrade} (inherits from \code{GeomRibbon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
GeomBalanceOfTrade
}
\description{
A ribbon between exports and imports, automatically coloring surplus vs deficit.
}
\keyword{datasets}
