% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-balance_of_trade.R
\docType{data}
\name{StatBalanceOfTrade}
\alias{StatBalanceOfTrade}
\title{StatBalanceOfTrade: split & interpolate exports–imports for ribbon plotting}
\format{
An object of class \code{StatBalanceOfTrade} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatBalanceOfTrade
}
\description{
(Transforms a data frame with aesthetics \code{x}, \code{exports}, and \code{imports}
into \code{ymin}, \code{ymax}, \code{direction}, and \code{group} for a ribbon.)
}
\keyword{datasets}
