% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-balance_of_trade.R
\name{geom_balance_of_trade}
\alias{geom_balance_of_trade}
\title{geom_balance_of_trade}
\usage{
geom_balance_of_trade(
  data = NULL,
  mapping = NULL,
  stat = StatBalanceOfTrade,
  geom = GeomBalanceOfTrade,
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA
)
}
\description{
Fill between exports and imports, coloring surplus vs deficit.
}
