% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-balance_of_trade.R
\docType{data}
\name{StatBalanceOfTrade}
\alias{StatBalanceOfTrade}
\title{StatBalanceOfTrade: split & interpolate export–import areas for ribbon plotting}
\format{
An object of class \code{StatBalanceOfTrade} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatBalanceOfTrade
}
\arguments{
\item{exports}{Numeric vector of export values}

\item{imports}{Numeric vector of import values}
}
\description{
StatBalanceOfTrade: split & interpolate export–import areas for ribbon plotting
}
\keyword{datasets}
